+++ src/main/java/com/coolerpromc/uncrafteverything/screen/custom/UEClientConfigScreen.java
@@ -17,6 +17,7 @@ import java.util.*;
 public class UEClientConfigScreen extends AbstractScrollableScreen {
     private final UncraftEverythingClientConfig config = UncraftEverythingClientConfig.CONFIG;
     public boolean autoMoveToInventory = config.autoMoveToInventory.getAsBoolean();
+    public boolean showJeiUncraftingCategory = config.showJeiUncraftingCategory.getAsBoolean();
     public int noRecipeFoundColor = config.noRecipeFoundColor.getAsInt();
     public int noSuitableOutputSlotColor = config.noSuitableOutputSlotColor.getAsInt();
     public int notEnoughExpColor = config.notEnoughExpColor.getAsInt();
@@ -32,32 +33,33 @@ public class UEClientConfigScreen extends AbstractScrollableScreen {
     private final Map<ColorPickerWidget, Tuple<Integer, Integer>> widgets = new HashMap<>();
 
     private Button moveToInventoryButton;
+    private Button showJeiCategoryButton;
     private Button cancelButton;
     private Button saveButton;
 
     public UEClientConfigScreen(Component title) {
-        super(title, 250);
-        ColorPickerWidget noRecipeFoundButton = new ColorPickerWidget(0, (int) (baseY + 25 - scrollAmount), 0, noRecipeFoundColor, Component.translatable("screen.uncrafteverything.config.no_recipe_found"), this, this::setNoRecipeFoundColor);
-        ColorPickerWidget noSuitableOutputSlotButton = new ColorPickerWidget(0, (int) (baseY + 50 - scrollAmount), 0, noSuitableOutputSlotColor, Component.translatable("screen.uncrafteverything.config.no_suitable_output_slot"), this, this::setNoSuitableOutputSlotColor);
-        ColorPickerWidget notEnoughExpButton = new ColorPickerWidget(0, (int) (baseY + 75 - scrollAmount), 0, notEnoughExpColor, Component.translatable("screen.uncrafteverything.config.not_enough_exp"), this, this::setNotEnoughExpColor);
-        ColorPickerWidget notEnoughInputItemButton = new ColorPickerWidget(0, (int) (baseY + 100 - scrollAmount), 0, notEnoughInputItemColor, Component.translatable("screen.uncrafteverything.config.not_enough_input_item"), this, this::setNotEnoughInputItemColor);
-        ColorPickerWidget notEmptyShulkerButton = new ColorPickerWidget(0, (int) (baseY + 125 - scrollAmount), 0, notEmptyShulkerColor, Component.translatable("screen.uncrafteverything.config.not_empty_shulker"), this, this::setNotEmptyShulkerColor);
-        ColorPickerWidget restrictedByConfigButton = new ColorPickerWidget(0, (int) (baseY + 150 - scrollAmount), 0, restrictedByConfigColor, Component.translatable("screen.uncrafteverything.config.restricted_by_config"), this, this::setRestrictedByConfigColor);
-        ColorPickerWidget damagedItemButton = new ColorPickerWidget(0, (int) (baseY + 175 - scrollAmount), 0, damagedItemColor, Component.translatable("screen.uncrafteverything.config.damaged_item"), this, this::setDamagedItemColor);
-        ColorPickerWidget enchantedItemButton = new ColorPickerWidget(0, (int) (baseY + 200 - scrollAmount), 0, enchantedItemColor, Component.translatable("screen.uncrafteverything.config.enchanted_item"), this, this::setEnchantedItemColor);
-        ColorPickerWidget lockedItemButton = new ColorPickerWidget(0, (int) (baseY + 225 - scrollAmount), 0, lockedItemColor, Component.translatable("screen.uncrafteverything.config.locked_item"), this, this::setLockedItemColor);
-        ColorPickerWidget progressionNotDefinedButton = new ColorPickerWidget(0, (int) (baseY + 250 - scrollAmount), 0, progressionNotDefinedColor, Component.translatable("screen.uncrafteverything.config.progression_not_defined"), this, this::setProgressionNotDefinedColor);
+        super(title, 275);
+        ColorPickerWidget noRecipeFoundButton = new ColorPickerWidget(0, (int) (baseY + 50 - scrollAmount), 0, noRecipeFoundColor, Component.translatable("screen.uncrafteverything.config.no_recipe_found"), this, this::setNoRecipeFoundColor);
+        ColorPickerWidget noSuitableOutputSlotButton = new ColorPickerWidget(0, (int) (baseY + 75 - scrollAmount), 0, noSuitableOutputSlotColor, Component.translatable("screen.uncrafteverything.config.no_suitable_output_slot"), this, this::setNoSuitableOutputSlotColor);
+        ColorPickerWidget notEnoughExpButton = new ColorPickerWidget(0, (int) (baseY + 100 - scrollAmount), 0, notEnoughExpColor, Component.translatable("screen.uncrafteverything.config.not_enough_exp"), this, this::setNotEnoughExpColor);
+        ColorPickerWidget notEnoughInputItemButton = new ColorPickerWidget(0, (int) (baseY + 125 - scrollAmount), 0, notEnoughInputItemColor, Component.translatable("screen.uncrafteverything.config.not_enough_input_item"), this, this::setNotEnoughInputItemColor);
+        ColorPickerWidget notEmptyShulkerButton = new ColorPickerWidget(0, (int) (baseY + 150 - scrollAmount), 0, notEmptyShulkerColor, Component.translatable("screen.uncrafteverything.config.not_empty_shulker"), this, this::setNotEmptyShulkerColor);
+        ColorPickerWidget restrictedByConfigButton = new ColorPickerWidget(0, (int) (baseY + 175 - scrollAmount), 0, restrictedByConfigColor, Component.translatable("screen.uncrafteverything.config.restricted_by_config"), this, this::setRestrictedByConfigColor);
+        ColorPickerWidget damagedItemButton = new ColorPickerWidget(0, (int) (baseY + 200 - scrollAmount), 0, damagedItemColor, Component.translatable("screen.uncrafteverything.config.damaged_item"), this, this::setDamagedItemColor);
+        ColorPickerWidget enchantedItemButton = new ColorPickerWidget(0, (int) (baseY + 225 - scrollAmount), 0, enchantedItemColor, Component.translatable("screen.uncrafteverything.config.enchanted_item"), this, this::setEnchantedItemColor);
+        ColorPickerWidget lockedItemButton = new ColorPickerWidget(0, (int) (baseY + 250 - scrollAmount), 0, lockedItemColor, Component.translatable("screen.uncrafteverything.config.locked_item"), this, this::setLockedItemColor);
+        ColorPickerWidget progressionNotDefinedButton = new ColorPickerWidget(0, (int) (baseY + 275 - scrollAmount), 0, progressionNotDefinedColor, Component.translatable("screen.uncrafteverything.config.progression_not_defined"), this, this::setProgressionNotDefinedColor);
 
-        widgets.put(noRecipeFoundButton, new Tuple<>(noRecipeFoundColor, 25));
-        widgets.put(noSuitableOutputSlotButton, new Tuple<>(noSuitableOutputSlotColor, 50));
-        widgets.put(notEnoughExpButton, new Tuple<>(notEnoughExpColor, 75));
-        widgets.put(notEnoughInputItemButton, new Tuple<>(notEnoughInputItemColor, 100));
-        widgets.put(notEmptyShulkerButton, new Tuple<>(notEmptyShulkerColor, 125));
-        widgets.put(restrictedByConfigButton, new Tuple<>(restrictedByConfigColor, 150));
-        widgets.put(damagedItemButton, new Tuple<>(damagedItemColor, 175));
-        widgets.put(enchantedItemButton, new Tuple<>(enchantedItemColor, 200));
-        widgets.put(lockedItemButton, new Tuple<>(lockedItemColor, 225));
-        widgets.put(progressionNotDefinedButton, new Tuple<>(progressionNotDefinedColor, 250));
+        widgets.put(noRecipeFoundButton, new Tuple<>(noRecipeFoundColor, 50));
+        widgets.put(noSuitableOutputSlotButton, new Tuple<>(noSuitableOutputSlotColor, 75));
+        widgets.put(notEnoughExpButton, new Tuple<>(notEnoughExpColor, 100));
+        widgets.put(notEnoughInputItemButton, new Tuple<>(notEnoughInputItemColor, 125));
+        widgets.put(notEmptyShulkerButton, new Tuple<>(notEmptyShulkerColor, 150));
+        widgets.put(restrictedByConfigButton, new Tuple<>(restrictedByConfigColor, 175));
+        widgets.put(damagedItemButton, new Tuple<>(damagedItemColor, 200));
+        widgets.put(enchantedItemButton, new Tuple<>(enchantedItemColor, 225));
+        widgets.put(lockedItemButton, new Tuple<>(lockedItemColor, 250));
+        widgets.put(progressionNotDefinedButton, new Tuple<>(progressionNotDefinedColor, 275));
     }
 
     @Override
@@ -70,6 +72,12 @@ public class UEClientConfigScreen extends AbstractScrollableScreen {
             button.setMessage(Component.translatable("screen.uncrafteverything.config.move_to_inventory_" + autoMoveToInventory));
         }).bounds(x, (int) (baseY - scrollAmount), widgetWidth, 20).build();
         this.addRenderableWidget(moveToInventoryButton);
+
+        showJeiCategoryButton = Button.builder(Component.translatable("screen.uncrafteverything.config.show_jei_category_" + showJeiUncraftingCategory), button -> {
+            showJeiUncraftingCategory = !showJeiUncraftingCategory;
+            button.setMessage(Component.translatable("screen.uncrafteverything.config.show_jei_category_" + showJeiUncraftingCategory));
+        }).bounds(x, (int) (baseY + 25 - scrollAmount), widgetWidth, 20).build();
+        this.addRenderableWidget(showJeiCategoryButton);
 
         Tuple<ColorPickerWidget, Tuple<Integer, Integer>> visible = null;
 
@@ -120,35 +128,38 @@ public class UEClientConfigScreen extends AbstractScrollableScreen {
         Component moveToInventory = Component.translatable("screen.uncrafteverything.config.move_to_inventory");
         pGuiGraphics.drawWordWrap(this.font, moveToInventory, x, (int) (baseY - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
+        Component showJeiCategory = Component.translatable("screen.uncrafteverything.config.show_jei_category");
+        pGuiGraphics.drawWordWrap(this.font, showJeiCategory, x, (int) (baseY + 25 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+
         Component noRecipeFound = Component.translatable("screen.uncrafteverything.config.no_recipe_found");
-        pGuiGraphics.drawWordWrap(this.font, noRecipeFound, x, (int) (baseY + 25 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.drawWordWrap(this.font, noRecipeFound, x, (int) (baseY + 50 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
         Component noSuitableOutputSlot = Component.translatable("screen.uncrafteverything.config.no_suitable_output_slot");
-        pGuiGraphics.drawWordWrap(this.font, noSuitableOutputSlot, x, (int) (baseY + 50 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.drawWordWrap(this.font, noSuitableOutputSlot, x, (int) (baseY + 75 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
         Component notEnoughExpSlot = Component.translatable("screen.uncrafteverything.config.not_enough_exp");
-        pGuiGraphics.drawWordWrap(this.font, notEnoughExpSlot, x, (int) (baseY + 75 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.drawWordWrap(this.font, notEnoughExpSlot, x, (int) (baseY + 100 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
         Component notEnoughInputItemSlot = Component.translatable("screen.uncrafteverything.config.not_enough_input_item");
-        pGuiGraphics.drawWordWrap(this.font, notEnoughInputItemSlot, x, (int) (baseY + 100 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.drawWordWrap(this.font, notEnoughInputItemSlot, x, (int) (baseY + 125 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
         Component notEmptyShulkerSlot = Component.translatable("screen.uncrafteverything.config.not_empty_shulker");
-        pGuiGraphics.drawWordWrap(this.font, notEmptyShulkerSlot, x, (int) (baseY + 125 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.drawWordWrap(this.font, notEmptyShulkerSlot, x, (int) (baseY + 150 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
         Component restrictedByConfigSlot = Component.translatable("screen.uncrafteverything.config.restricted_by_config");
-        pGuiGraphics.drawWordWrap(this.font, restrictedByConfigSlot, x, (int) (baseY + 150 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.drawWordWrap(this.font, restrictedByConfigSlot, x, (int) (baseY + 175 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
         Component damagedItemSlot = Component.translatable("screen.uncrafteverything.config.damaged_item");
-        pGuiGraphics.drawWordWrap(this.font, damagedItemSlot, x, (int) (baseY + 175 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.drawWordWrap(this.font, damagedItemSlot, x, (int) (baseY + 200 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
         Component enchantedItemSlot = Component.translatable("screen.uncrafteverything.config.enchanted_item");
-        pGuiGraphics.drawWordWrap(this.font, enchantedItemSlot, x, (int) (baseY + 200 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.drawWordWrap(this.font, enchantedItemSlot, x, (int) (baseY + 225 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
         Component lockedItemSlot = Component.translatable("screen.uncrafteverything.config.locked_item");
-        pGuiGraphics.drawWordWrap(this.font, lockedItemSlot, x, (int) (baseY + 225 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.drawWordWrap(this.font, lockedItemSlot, x, (int) (baseY + 250 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
         Component progressionNotDefinedSlot = Component.translatable("screen.uncrafteverything.config.progression_not_defined");
-        pGuiGraphics.drawWordWrap(this.font, progressionNotDefinedSlot, x, (int) (baseY + 250 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
+        pGuiGraphics.drawWordWrap(this.font, progressionNotDefinedSlot, x, (int) (baseY + 275 - scrollAmount + (this.font.lineHeight / 2d) + 2), textWidth, 0xFFFFFFFF);
 
         for (Renderable renderable : this.renderables) {
             renderable.render(pGuiGraphics, pMouseX, pMouseY, pPartialTick);
@@ -268,6 +279,7 @@ public class UEClientConfigScreen extends AbstractScrollableScreen {
 
     private void pressSaveButton(Button button){
         config.autoMoveToInventory.set(this.autoMoveToInventory);
+        config.showJeiUncraftingCategory.set(this.showJeiUncraftingCategory);
         config.noRecipeFoundColor.set(this.noRecipeFoundColor);
         config.noSuitableOutputSlotColor.set(this.noSuitableOutputSlotColor);
         config.notEnoughExpColor.set(this.notEnoughExpColor);
@@ -303,6 +315,16 @@ public class UEClientConfigScreen extends AbstractScrollableScreen {
             );
             guiGraphics.renderTooltip(this.font, tooltip, Optional.empty(), mouseX, mouseY);
         }
+
+        if (showJeiCategoryButton.isHovered()){
+            List<Component> tooltip = List.of(
+                    title("tooltip.uncrafteverything.config.show_jei_category"),
+                    valueInfo("tooltip.uncrafteverything.config.true", "tooltip.uncrafteverything.config.show_jei_category_true"),
+                    Component.empty(),
+                    valueInfo("tooltip.uncrafteverything.config.false", "tooltip.uncrafteverything.config.show_jei_category_false")
+            );
+            guiGraphics.renderTooltip(this.font, tooltip, Optional.empty(), mouseX, mouseY);
+        }
     }
 
     private Component title(String title){
@@ -319,51 +341,51 @@ public class UEClientConfigScreen extends AbstractScrollableScreen {
 
     private void setNoRecipeFoundColor(int noRecipeFoundColor, ColorPickerWidget button) {
         this.noRecipeFoundColor = noRecipeFoundColor;
-        this.widgets.put(button, new Tuple<>(noRecipeFoundColor, 25));
+        this.widgets.put(button, new Tuple<>(noRecipeFoundColor, 50));
     }
 
     private void setNoSuitableOutputSlotColor(int noSuitableOutputSlotColor, ColorPickerWidget button) {
         this.noSuitableOutputSlotColor = noSuitableOutputSlotColor;
-        this.widgets.put(button, new Tuple<>(noSuitableOutputSlotColor, 50));
+        this.widgets.put(button, new Tuple<>(noSuitableOutputSlotColor, 75));
     }
 
     public void setNotEnoughExpColor(int notEnoughExpColor, ColorPickerWidget button) {
         this.notEnoughExpColor = notEnoughExpColor;
-        this.widgets.put(button, new Tuple<>(notEnoughExpColor, 75));
+        this.widgets.put(button, new Tuple<>(notEnoughExpColor, 100));
     }
 
     public void setNotEnoughInputItemColor(int notEnoughInputItemColor, ColorPickerWidget button) {
         this.notEnoughInputItemColor = notEnoughInputItemColor;
-        this.widgets.put(button, new Tuple<>(notEnoughInputItemColor, 100));
+        this.widgets.put(button, new Tuple<>(notEnoughInputItemColor, 125));
     }
 
     public void setNotEmptyShulkerColor(int notEmptyShulkerColor, ColorPickerWidget button) {
         this.notEmptyShulkerColor = notEmptyShulkerColor;
-        this.widgets.put(button, new Tuple<>(notEmptyShulkerColor, 125));
+        this.widgets.put(button, new Tuple<>(notEmptyShulkerColor, 150));
     }
 
     public void setRestrictedByConfigColor(int restrictedByConfigColor, ColorPickerWidget button) {
         this.restrictedByConfigColor = restrictedByConfigColor;
-        this.widgets.put(button, new Tuple<>(restrictedByConfigColor, 150));
+        this.widgets.put(button, new Tuple<>(restrictedByConfigColor, 175));
     }
 
     public void setDamagedItemColor(int damagedItemColor, ColorPickerWidget button) {
         this.damagedItemColor = damagedItemColor;
-        this.widgets.put(button, new Tuple<>(damagedItemColor, 175));
+        this.widgets.put(button, new Tuple<>(damagedItemColor, 200));
     }
 
     public void setEnchantedItemColor(int enchantedItemColor, ColorPickerWidget button) {
         this.enchantedItemColor = enchantedItemColor;
-        this.widgets.put(button, new Tuple<>(enchantedItemColor, 200));
+        this.widgets.put(button, new Tuple<>(enchantedItemColor, 225));
     }
 
     public void setLockedItemColor(int lockedItemColor, ColorPickerWidget button) {
         this.lockedItemColor = lockedItemColor;
-        this.widgets.put(button, new Tuple<>(lockedItemColor, 225));
+        this.widgets.put(button, new Tuple<>(lockedItemColor, 250));
     }
 
     public void setProgressionNotDefinedColor(int progressionNotDefinedColor, ColorPickerWidget button) {
         this.progressionNotDefinedColor = progressionNotDefinedColor;
-        this.widgets.put(button, new Tuple<>(progressionNotDefinedColor, 250));
+        this.widgets.put(button, new Tuple<>(progressionNotDefinedColor, 275));
     }
 }